+++ test/test_integration.jl
@@ -1,6 +1,8 @@
 using DocumenterCitations
 using Documenter
+using Bijections
 using Test
+using TestingUtilities: @Test  # much better at comparing strings
 
 include("run_makedocs.jl")
 
@@ -82,9 +84,17 @@ end
     citations = OrderedDict{String,Int64}()
     page_citations = Dict{String,Set{String}}()
     anchor_map = Documenter.AnchorMap()
+    anchor_keys = Bijections.Bijection{String,String}()
 
-    bib =
-        CitationBibliography(bibfile, style, entries, citations, page_citations, anchor_map)
+    bib = CitationBibliography(
+        bibfile,
+        style,
+        entries,
+        citations,
+        page_citations,
+        anchor_map,
+        anchor_keys
+    )
 
     run_makedocs(
         joinpath(@__DIR__, "..", "docs");
@@ -100,7 +110,18 @@ end
         ],
         check_success=true
     ) do dir, result, success, backtrace, output
+
         @test success
+
+        # ArXiV references were (unnoticed) broken prior to PR #95
+
+        references_html = read(joinpath(dir, "build", "references", "index.html"), String)
+        @test !contains(references_html, "<div id=\"Wilhelm2003\">")  # pre-#95
+        @Test contains(references_html, "<div id=\"Wilhelm2003_10132\">")
+
+        syntax_html = read(joinpath(dir, "build", "syntax", "index.html"), String)
+        @test contains(syntax_html, "<a href=\"../references/#Wilhelm2003_10132\">")
+
     end
 
 end